fix: fix error type exports
